src/test/java/com/thoughtworks/bitemoi/RestaurantListAdapterTest.java
package com.thoughtworks.bitemoi;

import android.widget.BaseAdapter;
import com.xtremelabs.robolectric.Robolectric;
import com.xtremelabs.robolectric.RobolectricTestRunner;
import org.junit.Test;
import org.junit.runner.RunWith;

import java.util.ArrayList;
import java.util.List;

import static junit.framework.Assert.assertEquals;

@RunWith(RobolectricTestRunner.class)
public class RestaurantListAdapterTest {
    @Test
    public void testGetCount() throws Exception {
        List<Restaurant> restaurants = new ArrayList<Restaurant>();
        restaurants.add(new Restaurant(1, "soup company"));
        restaurants.add(new Restaurant(2, "cafe venue"));

        BaseAdapter adapter = new RestaurantListAdapter(Robolectric.application.getApplicationContext(), restaurants);

        assertEquals(adapter.getCount(), 2);

    }
}
